Improve color calculation, docs
