+++ .changeset/silent-carrots-divide.md
@@ -1,0 +1,5 @@
+---
+'better-color-tools': minor
+---
+
+Reorg library around Oklab colorspace
