+++ .eslintignore
@@ -1,1 +1,2 @@
 **/dist/**/*
+**/*.min.js
