+++ .prettierignore
@@ -1,3 +1,2 @@
 src/css-names.ts
 test/**/*.js
-/README.md
