+++ package.json
@@ -32,23 +32,24 @@
     "dev": "tsc -w",
     "lint": "eslint \"**/*.{js,ts}\"",
     "prepublishOnly": "npm run build",
-    "test": "mocha --parallel"
+    "test": "mocha --parallel",
+    "test:benchmark": "mocha"
   },
   "devDependencies": {
-    "@changesets/cli": "^2.20.0",
-    "@types/node": "^16.11.24",
-    "@typescript-eslint/eslint-plugin": "^5.11.0",
-    "@typescript-eslint/parser": "^5.11.0",
+    "@changesets/cli": "^2.21.0",
+    "@types/node": "^16.11.25",
+    "@typescript-eslint/eslint-plugin": "^5.12.1",
+    "@typescript-eslint/parser": "^5.12.1",
     "chai": "^4.3.6",
-    "esbuild": "^0.14.21",
+    "esbuild": "^0.14.23",
     "eslint": "^8.9.0",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint-config-prettier": "^8.4.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "mocha": "^9.2.0",
+    "mocha": "^9.2.1",
     "npm-run-all": "^4.1.5",
-    "number-precision": "^1.5.1",
+    "number-precision": "^1.5.2",
     "prettier": "^2.5.1",
-    "sass": "^1.49.7",
+    "sass": "^1.49.8",
     "typescript": "^4.5.5"
   }
 }
