--- src/hsl.ts
@@ -1,39 +1,0 @@
-export type HSL = [number, number, number, number]; // with alpha
-
-import type { sRGB } from './rgb.js';
-
-/** HSL -> sRGB */
-export function hslTosRGB(hsl: HSL): sRGB {
-  let [H, S, L, A] = hsl;
-  H = Math.abs(H % 360); // allow < 0 and > 360
-
-  const C = S * (1 - Math.abs(2 * L - 1));
-  const X = C * (1 - Math.abs(((H / 60) % 2) - 1));
-
-  let R = 0;
-  let G = 0;
-  let B = 0;
-
-  if (0 <= H && H < 60) {
-    R = C;
-    G = X;
-  } else if (60 <= H && H < 120) {
-    R = X;
-    G = C;
-  } else if (120 <= H && H < 180) {
-    G = C;
-    B = X;
-  } else if (180 <= H && H < 240) {
-    G = X;
-    B = C;
-  } else if (240 <= H && H < 300) {
-    R = X;
-    B = C;
-  } else if (300 <= H && H < 360) {
-    R = C;
-    B = X;
-  }
-  const m = L - C / 2;
-
-  return [R + m, G + m, B + m, A];
-}
