--- src/oklch.ts
@@ -1,1 +1,0 @@
-export type Oklch = [number, number, number, number];
