+++ www/README.md
@@ -1,0 +1,6 @@
+# www
+
+The website code for [better-color-tools.pages.dev](https://better-color-tools.pages.dev). The MIT license also applies to all unique code in here, with the obvious exception of other open source tools like [Astro](https://github.com/withastro/astro) which
+have their own licenses.
+
+Have a look around!
