+++ www/src/styles/_utils.scss
@@ -10,38 +10,82 @@
   font-weight: 300;
 }
 
-.color-block {
-  background-color: #242424;
+.color {
   color: white;
-  font-size: 14px;
-  height: 0;
-  padding-top: 100%;
-  position: relative;
-  transition: background-color 67ms linear, color 67ms linear;
-  width: 100%;
 
-  &.light {
-    color: rgba(0, 0, 0, 0.8);
+  &-block {
+    background-color: #242424;
+    display: flex;
+    flex: 1 0;
+    font-size: 14px;
+    overflow: hidden;
+    position: relative;
+    transition: background-color 67ms linear, color 67ms linear;
+
+    &-info {
+      box-sizing: border-box;
+      display: flex;
+      flex-direction: column;
+      font-family: var(--monospace);
+      font-weight: 300;
+      justify-content: flex-end;
+      line-height: 1.75;
+      min-height: 50%;
+      opacity: 0.8;
+      padding: 1.5rem 0 1.5rem 1.5rem;
+      white-space: nowrap;
+    }
+
+    &--s {
+      font-size: 12px;
+    }
   }
 
-  &.dark {
-    color: rgba(255, 255, 255, 0.85);
-  }
+  &-scale {
+    display: flex;
+    margin-bottom: 1px;
+    max-width: 100vw;
+    min-height: 10rem;
+    overflow: hidden;
+    position: relative;
 
-  &-info {
-    bottom: 0;
-    font-family: var(--monospace);
-    font-weight: 200;
-    height: 100%;
-    left: 0;
-    line-height: 1.75;
-    opacity: 0.8;
-    padding: 1.5rem;
-    position: absolute;
-    width: 100%;
-  }
+    &-title {
+      font-size: 24px;
+      font-weight: 450;
+      left: 1.5rem;
+      margin: 0;
+      position: absolute;
+      top: 1rem;
+      z-index: 5;
 
-  &--s {
-    font-size: 12px;
+      sup {
+        font-size: 0.5em;
+        left: 100%;
+        letter-spacing: 0.0125em;
+        margin-left: 0.25rem;
+        position: absolute;
+        text-transform: uppercase;
+        top: 0;
+      }
+    }
   }
+}
+
+.wrapper {
+  box-sizing: border-box;
+  width: 55rem;
+  padding: 1.5rem;
+  max-width: 100%;
+
+  @media (min-width: 650px) {
+    padding: 4rem;
+  }
+}
+
+.light {
+  color: rgba(0, 0, 0, 0.8);
+}
+
+.dark {
+  color: rgba(255, 255, 255, 0.85);
 }
